requirements.txt
importlib-metadata==7.0.1
setuptools==69.1.0
sqlparse==0.4.4
tabulate==0.9.0
wheel==0.42.0
yoyo-migrations==8.2.0
zipp==3.17.0
